Move modals to subdirectory
